console diffs working with colors
