+++ pax-compiler/src/design_server/mod.rs
@@ -27,7 +27,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use websocket::PrivilegedAgentWebSocket;
 use websocket::SocketMessageAccumulator;
 
-pub mod code_serialization;
 #[allow(unused)]
 mod llm;
 pub mod static_server;
