+++ pax-lang/Cargo.toml
@@ -16,6 +16,9 @@ pest_derive = {version = "2.7.10", features = ["std"]}
 serde = { version="1.0.95", features=["derive"]}
 pax-runtime-api = {path="../pax-runtime-api", version="0.36.9"}
 log = "0.4.20"
+syn = { version = "1.0", features = ["full", "parsing", "visit"] }
+color-eyre = "0.6.2"
+proc-macro2 = { version = "1.0", features = ["span-locations"] }
 
 [profile.dev]
 debug = false
