+++ pax-manifest/Cargo.toml
@@ -14,10 +14,18 @@ parsing = []
 [lib]
 
 [dependencies]
-serde = { version="1.0.95", features=["derive"]}
 pax-message = {path="../pax-message", version="0.36.9"}
 pax-runtime-api = {path="../pax-runtime-api", version="0.36.9"}
 pax-lang = {version = "0.36.9", path="../pax-lang"}
 kurbo = "0.9.0"
-serde_with = { version= "3.6.1", features = ["json"]} 
 log = "0.4.20"
+colored = "2.0.0"
+include_dir = {version = "0.7.3", features = ["glob"]}
+similar = "2.2.1"
+serde = { version = "1.0.159", features=["derive"] }
+serde_derive = "1.0.159"
+serde_with = { version= "3.6.1", features = ["json"]} 
+serde_json = { version = "1.0.95"}
+syn = { version = "1.0", features = ["full", "parsing", "visit"] }
+tera = "1"
+html-escape = "0.2.13"
