+++ pax-manifest/templates/code_serialization/macros.tera
