+++ pax-manifest/templates/code_serialization/manifest-code-serialization.tera
