+++ pax-manifest/templates/code_serialization/rust-file-serialization.tera
